requirements.txt
fastapi
uvicorn
streamlit
torch
torchvision
ultralytics
opencv-python
pillow
python-multipart
albumentations
scikit-learn
tqdm
plotly
folium
streamlit-folium
requests
